refactor: cambios en estructura de carpetas y nombres de archivos

+++ src/main/java/app/PetShere/repositories/UserRepository.java
@@ -1,9 +1,9 @@
 package app.PetShere.repositories;
 
-import app.PetShere.models.User;
+import app.PetShere.models.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
 }
